// apps/api-server/src/interfaces/mongoose.gen.ts
export type UserQueries = {};

export type UserMethods = {
  comparePassword: (this: UserDocument, ...args: any[]) => any;
  generateAuthTokens: (this: UserDocument, ...args: any[]) => any;
  generateAccessToken: (this: UserDocument, ...args: any[]) => any;
  me: (this: UserDocument, ...args: any[]) => any;
};

export type UserStatics = {};

/**
 * Mongoose Model type
 *
 * Pass this type to the Mongoose Model constructor:
 * ```
 * const User = mongoose.model<UserDocument, UserModel>("User", UserSchema);
 * ```
 */
export type UserModel = mongoose.Model<UserDocument, UserQueries> & UserStatics;

/**
 * Mongoose Schema type
 *
 * Assign this type to new User schema instances:
 * ```
 * const UserSchema: UserSchema = new mongoose.Schema({ ... })
 * ```
 */
export type UserSchema = mongoose.Schema<
  UserDocument,
  UserModel,
  UserMethods,
  UserQueries
>;

/**
 * Mongoose Subdocument type
 *
 * Type of `UserDocument["devices"]` element.
 */
export type UserDeviceDocument = mongoose.Types.Subdocument & {
  ip: string;
  geo: any;
  clientInfo: any;
  token: number;
  _id: mongoose.Types.ObjectId;
  updatedAt?: Date;
  createdAt?: Date;
};

/**
 * Mongoose Document type
 *
 * Pass this type to the Mongoose Model constructor:
 * ```
 * const User = mongoose.model<UserDocument, UserModel>("User", UserSchema);
 * ```
 */
export type UserDocument = mongoose.Document<
  mongoose.Types.ObjectId,
  UserQueries
> &
  UserMethods & {
    name: string;
    email: string;
    username: string;
    avatar?: string;
    bio?: string;
    password: string;
    isEmailVerified?: boolean;
    otp?: string;
    otpExpiry?: number;
    secondaryEmail?: string;
    isSecondaryEmailVerified?: boolean;
    secondaryEmailOtp?: string;
    secondaryEmailOtpExpiry?: number;
    tokenVersion: number;
    mfaTypes: mongoose.Types.Array<"emailOtp" | "authenticator">;
    mfaTempSecret?: any;
    mfaSecret?: any;
    mfaOtp?: string;
    mfaOtpExpiry?: number;
    _id: mongoose.Types.ObjectId;
    updatedAt?: Date;
    createdAt?: Date;
    devices: mongoose.Types.DocumentArray<UserDeviceDocument>;
  };